Type uint64 fields as uint64 instead of int
Fields that are uint64 upstream but were typed int (= number) can silently be
a BigInt at runtime: the RPC JSON parser returns BigInt for integers above
2^53-1 (~9,007 XCH in mojos). Retype the 28 affected fields to uint64
(= number | bigint) so amounts, fees, costs, byte sizes and timestamps are
sound.

+++ src/api/chia/wallet/trade_record.ts
@@ -36,7 +36,7 @@ export type TradeRecordConvenience = {
     offered: Record<str, int>; // {[asset_id]: amount}
     requested: Record<str, int>; // {[asset_id]: amount}
     infos: TDriverDict;
-    fees: int;
+    fees: uint64;
   };
   pending: Record<str, int>; // {[asset_id]: amount}
 } & Omit<TradeRecord, "offer">;
